intranet_api/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
deep-translator==1.11.1
Django==4.1.5
django-auth-ldap==4.3.0
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
langdetect==1.0.9
ldap3==2.9.1
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-ldap==3.4.3
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
